Finally fixed freezing GUIs caused by progress dialogs.
A number of other smaller things.

+++ ffmpeg.py
@@ -19,14 +19,16 @@
 from __future__ import division
 
 import glob
+import itertools
 import logging
 import os
-import sys
 import subprocess
+import sys
 import threading
 
 import requests
 import wx
+import wx.lib.intctrl as intctrl
 
 import seven_zip
 from jam_tools import wrap_exceptions
@@ -38,32 +40,51 @@ except ImportError:
 
 ENCODER_START = 36
 ENCODER_LEN = 34
+PD_STYLE = wx.PD_APP_MODAL | wx.PD_AUTO_HIDE | wx.PD_CAN_ABORT | wx.PD_ELAPSED_TIME | wx.PD_ESTIMATED_TIME
 
+# Converted from a list of MIME types. If there is something missing, file an issue or create a PR!
+FILE_EXTS = (
+    '*.3g2', '*.3gp', '*.aac', '*.adp', '*.aif', '*.aifc', '*.aiff', '*.asf', '*.asx', '*.au', '*.avi', '*.caf',
+    '*.dra', '*.dts', '*.dtshd', '*.dvb', '*.ecelp4800', '*.ecelp7470', '*.ecelp9600', '*.eol', '*.f4v', '*.flac',
+    '*.fli', '*.flv', '*.fvt', '*.h261', '*.h263', '*.h264', '*.jpgm', '*.jpgv', '*.jpm', '*.kar', '*.lvp',
+    '*.m1v', '*.m2a', '*.m2v', '*.m3a', '*.m3u', '*.m4a', '*.m4u', '*.m4v', '*.mid', '*.midi', '*.mj2', '*.mjp2',
+    '*.mk3d', '*.mka', '*.mks', '*.mkv', '*.mng', '*.mov', '*.movie', '*.mp2', '*.mp2a', '*.mp3', '*.mp4',
+    '*.mp4a', '*.mp4v', '*.mpa', '*.mpe', '*.mpeg', '*.mpg', '*.mpg4', '*.mpga', '*.mxu', '*.oga', '*.ogg',
+    '*.ogv', '*.pya', '*.pyv', '*.qt', '*.ra', '*.ram', '*.rip', '*.rmi', '*.rmp', '*.s3m', '*.sil', '*.smv',
+    '*.snd', '*.spx', '*.uva', '*.uvh', '*.uvm', '*.uvp', '*.uvs', '*.uvu', '*.uvv', '*.uvva', '*.uvvh', '*.uvvm',
+    '*.uvvp', '*.uvvs', '*.uvvu', '*.uvvv', '*.viv', '*.vob', '*.wav', '*.wax', '*.weba', '*.webm', '*.wm',
+    '*.wma', '*.wmv', '*.wmx', '*.wvx', '*.xm'
+)
 
 
 class FFmpegDownloaderThread(threading.Thread):
     def __init__(self, parent, url, file_size):
         super(FFmpegDownloaderThread, self).__init__()
-        self.parent = parent
         self.url = url
+        self.parent = parent
         self.file_size = file_size
-        self.running = True
+        self._abort = threading.Event()
+
         self.daemon = True
         self.start()
 
     def abort(self):
-        self.running = False
+        self._abort.set()
+
+    def is_aborted(self):
+        return self._abort.isSet()
 
     @wrap_exceptions
     def run(self):
         file_size_dl = 0
         response = requests.get(self.url, stream=True)
-        # chunk_size=1024 used to be fairly quick when testing, but now it's slow as a snail. weird. 1 MB chunks instead
-        data_chunks = response.iter_content(chunk_size=1024 ** 2)
+        data_chunks = response.iter_content(chunk_size=1024)
+
         if not os.path.exists('bin'):
             os.mkdir('bin')
+
         with open('bin/ffmpeg.7z', 'wb') as f:
-            while self.running:
+            while not self.is_aborted():
                 try:
                     chunk = next(data_chunks)
                     file_size_dl += len(chunk)
@@ -71,7 +92,7 @@ class FFmpegDownloaderThread(threading.Thread):
                         f.write(chunk)
                         f.flush()
                         os.fsync(f.fileno())
-                    wx.CallAfter(self.parent.update, file_size_dl)
+                    wx.CallAfter(self.parent.update, message=file_size_dl)
                 except StopIteration:
                     wx.CallAfter(self.parent.complete)
                     break
@@ -80,11 +101,10 @@ class FFmpegDownloaderThread(threading.Thread):
 class FFmpegDownloader(wx.ProgressDialog):
     def __init__(self, parent, url):
         file_size = int(requests.head(url).headers["Content-Length"])
-        super(FFmpegDownloader, self).__init__("Download in progress", "Downloading FFmpeg...", file_size // 1024,
-                                               parent, style=wx.PD_APP_MODAL | wx.PD_ELAPSED_TIME |
-                                                             wx.PD_REMAINING_TIME | wx.PD_CAN_ABORT)
+        super(FFmpegDownloader, self).__init__(title="Download in progress", message="Downloading FFmpeg...",
+                                               maximum=file_size // 1024, parent=parent, style=PD_STYLE)
 
-        self.downloader = FFmpegDownloaderThread(self, url, file_size)
+        self.downloader = FFmpegDownloaderThread(parent=self, url=url, file_size=file_size)
 
     def update(self, message):
         if self:  # True PD has not been destroyed.
@@ -93,25 +113,28 @@ class FFmpegDownloader(wx.ProgressDialog):
                 self.downloader.abort()
                 logging.info("FFmpeg download canceled.")
 
-                alert = wx.MessageDialog(self, "Aborted! FFmpeg was not downloaded.", "pyjam", wx.ICON_EXCLAMATION)
+                alert = wx.MessageDialog(parent=self, message="Aborted! FFmpeg was not downloaded.", caption="pyjam",
+                                         style=wx.ICON_EXCLAMATION)
                 alert.ShowModal()
                 alert.Destroy()
 
                 wx.CallAfter(self.Destroy)
 
     def complete(self):
-        logging.info("FFmpeg download complete.")
-        if seven_zip.find() is None:
-            message_str = "FFmpeg was downloaded succesfully!\nPlease extract it. (bin/ffmpeg.7z)"
-        else:
-            seven_zip.extract_single(os.path.normpath('bin/ffmpeg.7z'), 'ffmpeg.exe', os.path.normpath('bin/'))
-            os.remove(os.path.normpath('bin/ffmpeg.7z'))
-            message_str = "FFmpeg was downloaded succesfully!"
+        if self:
+            logging.info("FFmpeg download complete.")
 
-        message = wx.MessageDialog(self, message_str, "pyjam")
-        message.ShowModal()
-        message.Destroy()
-        wx.CallAfter(self.Destroy)
+            if seven_zip.find() is None:
+                message_str = "FFmpeg was downloaded succesfully!\nPlease extract it. (bin/ffmpeg.7z)"
+            else:
+                seven_zip.extract_single(os.path.normpath('bin/ffmpeg.7z'), 'ffmpeg.exe', os.path.normpath('bin/'))
+                os.remove(os.path.normpath('bin/ffmpeg.7z'))
+                message_str = "FFmpeg was downloaded succesfully!"
+            message = wx.MessageDialog(self, message_str, "pyjam")
+            message.ShowModal()
+            message.Destroy()
+
+            wx.CallAfter(self.Destroy)
 
 
 class FFmpegConvertThread(threading.Thread):
@@ -123,18 +146,22 @@ class FFmpegConvertThread(threading.Thread):
         self.vol = vol
         self.songs = songs
         self.converted = 0
-        self.running = True
+        self._running = threading.Event()
+
         self.daemon = True
         self.start()
 
     def abort(self):
-        self.running = False
+        self._running.set()
+
+    def is_aborted(self):
+        return self._running.isSet()
 
     @wrap_exceptions
     def run(self):
         tracks = iter(self.songs)
-        errors = 0
-        while self.running:
+        errors = []
+        while not self.is_aborted():
             try:
                 track = next(tracks)
                 file = os.path.join(self.dest, os.path.splitext(os.path.basename(track))[0])
@@ -143,44 +170,50 @@ class FFmpegConvertThread(threading.Thread):
                                                                                                 vol=self.vol))
                 convert = convert_audio(track, file, self.rate, self.vol)
                 if convert != 0:
-                    errors += 1
+                    errors.append(track)
+                    # File's headers aren't stripped, which normally would increase self.converted by 1.
+                    self.converted += 1
                 else:
                     self.strip_encoder(file)
                 self.converted += 1
-                self.parent.update(self.converted)
+                wx.CallAfter(self.parent.update, message=self.converted)
             except StopIteration:
-                self.parent.complete(errors)
+                wx.CallAfter(self.parent.complete, message=errors)
                 break
 
     def strip_encoder(self, file):
-        logger.info("Stripping metadata from {file}".format(file=file))
+        logger.info("Stripping metadata from {file}".format(file=file + '.wav'))
         with open(file + '.wav', 'rb') as f:
             wav = bytearray(f.read())
         del wav[ENCODER_START:ENCODER_START + ENCODER_LEN]
         with open(file + '.wav', 'wb') as f:
             f.write(wav)
         self.converted += 1
-        self.parent.update(self.converted)
+        wx.CallAfter(self.parent.update, message=self.converted)
+
 
 class FFmpegConvertDialog(wx.Dialog):
-    def __init__(self, parent, rate, out):
-        super(FFmpegConvertDialog, self).__init__(parent, title="Audio Converter",
+    def __init__(self, parent, rate=None, out_dir=None, in_dir=None):
+        super(FFmpegConvertDialog, self).__init__(parent=parent, title="pyjam Audio Converter",
                                                   style=wx.DEFAULT_DIALOG_STYLE | wx.RESIZE_BORDER)
-        self.parent = parent
 
-        self.game_rate = wx.TextCtrl(self)
+        self.game_rate = intctrl.IntCtrl(self)
         game_rate_text = wx.StaticText(self, label="Audio rate")
-        self.game_rate.SetValue(rate)
+        if rate is not None:
+            self.game_rate.SetValue(rate)
 
         self.volume = wx.SpinCtrl(self, initial=85)
         volume_text = wx.StaticText(self, label="Volume %")
 
         self.in_dir = wx.DirPickerCtrl(self, name="Input directory")
         in_dir_text = wx.StaticText(self, label="Input directory")
+        if in_dir is not None:
+            self.in_dir.SetPath(in_dir)
 
         self.out_dir = wx.DirPickerCtrl(self, name="Output directory")
         out_dir_text = wx.StaticText(self, label="Output directory")
-        self.out_dir.SetPath(out)
+        if out_dir is not None:
+            self.out_dir.SetPath(out_dir)
 
         top_sizer = wx.BoxSizer(wx.VERTICAL)
         top_row = wx.BoxSizer(wx.HORIZONTAL)
@@ -201,7 +234,7 @@ class FFmpegConvertDialog(wx.Dialog):
         dir_sizer.Add(self.out_dir, 0, wx.ALL ^ wx.LEFT ^ wx.TOP | wx.ALIGN_LEFT | wx.EXPAND, 5)
         top_sizer.Add(top_row)
         top_sizer.Add(dir_sizer, 1, wx.ALL | wx.EXPAND | wx.ALIGN_TOP, 5)
-        top_sizer.Add(button_sizer, 0, wx.ALL | wx.CENTER, 5)
+        top_sizer.Add(button_sizer, 0, wx.ALL | wx.ALIGN_CENTER, 5)
 
         self.Bind(wx.EVT_BUTTON, self.on_ok, id=wx.ID_OK)
 
@@ -214,41 +247,58 @@ class FFmpegConvertDialog(wx.Dialog):
         self.ShowModal()
 
     def on_ok(self, event):
-        songs = glob.glob(os.path.join(self.in_dir.GetPath(), '*'))
+        songs = list(itertools.chain.from_iterable((glob.iglob(os.path.join(self.in_dir.GetPath(), x))
+                                                    for x in FILE_EXTS)))
         self.num_songs = len(songs)
         if self.num_songs <= 0:
-            alert = wx.MessageDialog(self, "No files in directory!", "pyjam", wx.ICON_EXCLAMATION)
+            alert = wx.MessageDialog(self, "No compatible files in directory!", "pyjam", wx.ICON_EXCLAMATION)
             alert.ShowModal()
             alert.Destroy()
             return
 
-        self.progress_dialog = wx.ProgressDialog("Conversion", "Converting songs...", self.num_songs * 2, parent=self,
-                                                 style=wx.PD_ELAPSED_TIME | wx.PD_CAN_ABORT | wx.PD_AUTO_HIDE | wx.PD_APP_MODAL)
+        self.progress_dialog = wx.ProgressDialog(title="Conversion", message="Converting songs...",
+                                                 maximum=self.num_songs * 2, parent=self, style=PD_STYLE)
 
-        self.converter = FFmpegConvertThread(self, self.out_dir.GetPath(), self.game_rate.GetValue(),
-                                             self.volume.GetValue(), songs)
+        self.converter = FFmpegConvertThread(parent=self, dest=self.out_dir.GetPath(), rate=self.game_rate.GetValue(),
+                                             vol=self.volume.GetValue(), songs=songs)
 
     def update(self, message):
-        progress = "{songs} out of {total}".format(songs=message, total=self.num_songs)
-        if self.progress_dialog:
+        progress = "{songs} out of {total}".format(songs=message // 2, total=self.num_songs)
+        if self.progress_dialog and self.converter.isAlive():
             if not self.progress_dialog.Update(value=message, newmsg="Converted: {prog}".format(prog=progress))[0]:
                 self.converter.abort()
+                self.converter.join()
+                self.progress_dialog.Destroy()
 
                 alert_string = "Aborted! Only {progress} songs were converted".format(progress=progress)
-                alert = wx.MessageDialog(self.progress_dialog, alert_string, "pyjam", wx.ICON_EXCLAMATION)
+                alert = wx.MessageDialog(parent=self, message=alert_string, caption="pyjam", style=wx.ICON_EXCLAMATION)
+                alert.ToggleWindowStyle(wx.STAY_ON_TOP)
                 alert.ShowModal()
                 alert.Destroy()
 
                 logging.info("Audio conversion canceled canceled.")
                 logging.info(progress)
-                wx.CallAfter(self.progress_dialog.Destroy)
+                # wx.CallAfter(self.progress_dialog.Destroy)
 
     def complete(self, message):
         if self.progress_dialog:
-            done_string = "Songs converted with {errors} error(s)".format(errors=message)
-            done_message = wx.MessageDialog(self.progress_dialog, done_string, "pyjam")
+            self.converter.join()
+            if len(message) != 0:
+                done_string = "Songs converted with {errors} error(s)".format(errors=len(message))
+            else:
+                done_string = "All songs were converted succesfully!"
+            done_message = wx.MessageDialog(parent=self, message=done_string, caption="pyjam")
+            done_message.ToggleWindowStyle(wx.STAY_ON_TOP)
             done_message.ShowModal()
             done_message.Destroy()
+
+            if len(message) != 0:
+                errors = '\n'.join(message)
+                error_dialog = wx.MessageDialog(parent=self, message="The following files caused errors\n" + errors,
+                                                caption="Conversion Error!", style=wx.OK | wx.ICON_ERROR)
+                error_dialog.ShowModal()
+                error_dialog.Destroy()
+                logger.critical("Error converting these files\n{errors}".format(errors=errors))
 
             logging.info(done_string)
             wx.CallAfter(self.progress_dialog.Destroy)
@@ -265,12 +315,9 @@ def find():
 
 
 def convert_audio(file, dest, rate, vol, codec="pcm_s16le"):
-    # type: (str, str, str, str, str) -> None
+    # type: (str, any, any, any, any) -> int
     cmd = '{ff} -y -i "{i}" -map_metadata -1 -ac 1 -aq 100 -acodec {codec} -ar {rate} -af volume={vol} "{dest}.wav"'
     cmd = cmd.format(ff=find(), i=file, codec=codec, rate=rate, vol=vol / 100, dest=dest)
     return subprocess.call(cmd)
-
-
-FILE_EXTS = ('*.3gp', '*.aac', '*.flv', '*.m4a', '*.mp3', '*.mp4', '*.ogg', '*.wav', '*.webm', '*.flac', '*.mkv')
 
 logger = logging.getLogger('jam.ffmpeg')
